requirements.txt
alembic==1.6.5
appdirs==1.4.4
asgiref==3.3.4
black==21.6b0
certifi==2021.5.30
click==8.0.1
colorama==0.4.4
distlib==0.3.2
fastapi==0.65.2
FastAPI-SQLAlchemy==0.2.1
filelock==3.0.12
greenlet==1.1.0
h11==0.12.0
Mako==1.1.4
MarkupSafe==2.0.1
mypy-extensions==0.4.3
pathspec==0.8.1
pipenv==2021.5.29
psycopg2==2.9.1
pydantic==1.8.2
python-dateutil==2.8.1
python-dotenv==0.18.0
python-editor==1.0.4
regex==2021.4.4
six==1.16.0
SQLAlchemy==1.4.19
starlette==0.14.2
toml==0.10.2
typing-extensions==3.10.0.0
uvicorn==0.14.0
virtualenv==20.4.7
virtualenv-clone==0.5.4
email-validator==1.1.3
python-jose==3.3.0
passlib==1.7.4
emails==0.6
python-multipart==0.0.5
boto3
itsdangerous==2.0.1
